New filters 'toc_only', 'inject_anchors'
The new filters are split-up versons of the existing `toc` filter.

`toc_only`, as the name states, returns only the `ul`-container with the
TOC.

`inject_anchors`, as the name states, returns the content's HTML with
TOC anchors injected.

The existing `toc` filters is just a concatenation of the outputs of
`toc` and `inject_anchors`.

And again some refactoring.

Signed-off-by: Torbjörn Klatt <opensource@torbjoern-klatt.de>